LPS-78886 user-associated-data-web - casts long to int

+++ modules/apps/foundation/user-associated-data/user-associated-data-web/src/main/java/com/liferay/user/associated/data/web/internal/portlet/action/ViewUADEntitiesMVCRenderCommand.java
@@ -163,7 +163,8 @@ public class ViewUADEntitiesMVCRenderCommand implements MVCRenderCommand {
 				selectedUserId, searchContainer.getStart(),
 				searchContainer.getEnd()));
 
-		searchContainer.setTotal(uadEntityAggregator.count(selectedUserId));
+		searchContainer.setTotal(
+			(int)uadEntityAggregator.count(selectedUserId));
 
 		return searchContainer;
 	}
